Use int, not double, in CprDecoder (when possible)

+++ src/ch/epfl/javions/adsb/CprDecoder.java
@@ -9,8 +9,8 @@ public final class CprDecoder {
     private static final int CPR_BITS = 17;
     private static final double CPR_ONE_HALF = scalb(0.5d, CPR_BITS);
 
-    private static final double LAT_ZONES_E = 60;
-    private static final double LAT_ZONES_O = LAT_ZONES_E - 1;
+    private static final int LAT_ZONES_E = 60;
+    private static final int LAT_ZONES_O = LAT_ZONES_E - 1;
 
     private static final double D_LAT_E = 1d / LAT_ZONES_E;
     private static final double D_LAT_O = 1d / LAT_ZONES_O;
@@ -28,7 +28,7 @@ public final class CprDecoder {
                                         int lonCprO,
                                         int latCprO,
                                         boolean mostRecentIsE) {
-        var latZIn = floor(scalb(LAT_ZONES_O * latCprE - LAT_ZONES_E * latCprO + CPR_ONE_HALF, -CPR_BITS));
+        var latZIn = (int) floor(scalb(LAT_ZONES_O * latCprE - LAT_ZONES_E * latCprO + CPR_ONE_HALF, -CPR_BITS));
         var latE = D_LAT_E * (normalizeZoneIndex(latZIn, LAT_ZONES_E) + scalb(latCprE, -CPR_BITS));
         var latO = D_LAT_O * (normalizeZoneIndex(latZIn, LAT_ZONES_O) + scalb(latCprO, -CPR_BITS));
 
@@ -41,14 +41,14 @@ public final class CprDecoder {
                     : geoPos(scalb(lonCprO, -CPR_BITS), latO);
         } else {
             var lonZonesO = lonZonesE - 1;
-            var lonZIn = floor(scalb(lonZonesO * lonCprE - lonZonesE * lonCprO + CPR_ONE_HALF, -CPR_BITS));
+            var lonZIn = (int) floor(scalb(lonZonesO * lonCprE - lonZonesE * lonCprO + CPR_ONE_HALF, -CPR_BITS));
             return mostRecentIsE
                     ? geoPos(normalizeZoneIndex(lonZIn, lonZonesE) + scalb(lonCprE, -CPR_BITS) / lonZonesE, latE)
                     : geoPos(normalizeZoneIndex(lonZIn, lonZonesO) + scalb(lonCprO, -CPR_BITS) / lonZonesO, latO);
         }
     }
 
-    private static double normalizeZoneIndex(double zIn, double zonesCount) {
+    private static int normalizeZoneIndex(int zIn, int zonesCount) {
         return zIn < 0 ? zIn + zonesCount : zIn;
     }
